Moves the OverviewModal and all its components to a new file, adds some more types for plans, adds some error checking on loading a new plan, but that was probably useless if we are redoing it
