+++ frontend/types/plan.ts
@@ -1,9 +1,18 @@
-export type Plan = {
+export interface Plan {
   name: string,
   degree: string,
   percentage: number,
   startYear: number,
   startSem: number
   endYear: number,
-  id: string
+  id: string,
+  courses: PlannedCourses,
 }
+
+export interface PlannedCourses {
+  [year: string]: {
+    sem: CourseData[];
+  }
+};
+
+export type CourseData = [string, number] // [courseCode, grid_pos]
